+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(whensxmas)
-
-test_check("whensxmas")